rename calculatorUI -> CalculatorUI
